+++ src/app/api/graphql/types/companionRequest.ts
@@ -1,0 +1,20 @@
+import { gql } from "graphql-tag";
+
+export const companionRequestTypeDefs = gql`
+    enum FriendRequestStatus {
+        pending
+        accepted
+        rejected
+        blocked
+    }
+
+    type CompanionRequest {
+        id: ID!
+        sender: User!
+        receiver: User!
+        status: FriendRequestStatus!
+        createdAt: String!
+        updatedAt: String!
+    }
+`;
+export default companionRequestTypeDefs;
