+++ src/app/api/graphql/types/user.ts
@@ -1,33 +1,6 @@
 import { gql } from "graphql-tag";
 
 export const userTypeDefs = gql`
-    type Query {
-        findUsers(email: String, name: String): [User!]!
-        getApplications(userId: String): [CompanionRequest]
-    }
-
-    type Mutation {
-        companionRequest(senderId: String!, receiverId: String!): CompanionRequest!
-        acceptFriendRequest(requestId: String!): CompanionRequest!
-        rejectFriendRequest(requestId: String!): CompanionRequest!
-    }
-
-    enum FriendRequestStatus {
-        pending
-        accepted
-        rejected
-        blocked
-    }
-
-    type CompanionRequest {
-        id: ID!
-        sender: User!
-        receiver: User!
-        status: FriendRequestStatus!
-        createdAt: String!
-        updatedAt: String!
-    }
-
     type User {
         _id: ID!
         name: String
